Coin Change Problem
We are working at the cash counter at a fun-fair, and  we have different types of coins available to  us in infinite quantities. The value of each coin is already given. Here we have to determine the number of ways of making change for a particular number of units using the given types of coins